ACE1 compat - two more tests
